easyocr-backend/requirements.txt
flask
flask-cors
easyocr
pillow
torch
